[Feat] 관심축제 등록 api 추가
